requirements.txt
streamlit
google-generativeai
numpy
pyyaml
python-dotenv
